Bullhorn done

+++ src/customTools/DbUser.java
@@ -20,7 +20,8 @@ public class DbUser {
  * @param userID - primary key from database. Must be type long
  * @return Bhuser
  */
-	public static Bhuser getUser(long userID)
+	//public static Bhuser getUser(long userID)
+	public static Bhuser getUser(int userID)
 	{
 		EntityManager em = DbUtil.getEmFactory().createEntityManager();
 		Bhuser user = em.find(Bhuser.class, userID);
